requirements.txt
torch
torchvision
numpy
torchsummary
tqdm
Pillow